refactor(web): adopt canonical PresignedUploadResult
