+++ packages/contracts/src/v1/listings/dto/ad.ts
@@ -6,3 +6,11 @@ import type { Ad as SchemaAd } from '../schema/ad.schema';
  */
 export type AdLocation = SchemaAd["location"];
 
+export interface PresignedUploadResult {
+    uploadUrl: string;
+    publicUrl: string;
+    key: string;
+    expiresIn: number;
+}
+
+
